+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 * `push`-command:
   * Provide more detailed error message if creation of remote file fails.
+* Add `details` config option.
+  * If true, act as if `--details` is given on command line.
 
 ## v0.7.5 (2021-05-25)